Refactor document mapper to use field mappers
