+++ lib/Container/Compiler/DocumentFieldMapperPass/Block.php
@@ -17,6 +17,6 @@ use EzSystems\EzPlatformSolrSearchEngine\Container\Compiler\BaseDocumentFieldMap
  */
 class Block extends BaseDocumentFieldMapperPass
 {
-    const AGGREGATE_MAPPER_SERVICE_ID = 'ezpublish.search.solr.document_mapper.plugin.block';
+    const AGGREGATE_MAPPER_SERVICE_ID = 'ezpublish.search.solr.document_field_mapper.block';
     const AGGREGATE_MAPPER_SERVICE_TAG = self::AGGREGATE_MAPPER_SERVICE_ID;
 }
