+++ lib/Container/Compiler/DocumentFieldMapperPass/Content.php
@@ -17,6 +17,6 @@ use EzSystems\EzPlatformSolrSearchEngine\Container\Compiler\BaseDocumentFieldMap
  */
 class Content extends BaseDocumentFieldMapperPass
 {
-    const AGGREGATE_MAPPER_SERVICE_ID = 'ezpublish.search.solr.document_mapper.plugin.content';
+    const AGGREGATE_MAPPER_SERVICE_ID = 'ezpublish.search.solr.document_field_mapper.content';
     const AGGREGATE_MAPPER_SERVICE_TAG = self::AGGREGATE_MAPPER_SERVICE_ID;
 }
