+++ lib/Resources/config/container/solr.yml
@@ -1,5 +1,6 @@
 imports:
     - {resource: solr/criterion_visitors.yml}
+    - {resource: solr/document_field_mappers.yml}
     - {resource: solr/facet_builder_visitors.yml}
     - {resource: solr/field_value_mappers.yml}
     - {resource: solr/services.yml}
@@ -53,13 +54,7 @@ services:
             - @ezpublish.search.solr.document_field_mapper.content_translation
             - @ezpublish.search.solr.document_field_mapper.location
             - @ezpublish.search.solr.document_field_mapper.location_translation
-            - @ezpublish.search.common.field_registry
-            - @ezpublish.spi.persistence.content_handler
-            - @ezpublish.spi.persistence.location_handler
-            - @ezpublish.spi.persistence.content_type_handler
-            - @ezpublish.spi.persistence.object_state_handler
-            - @ezpublish.spi.persistence.section_handler
-            - @ezpublish.search.common.field_name_generator
+            - '@ezpublish.spi.persistence.location_handler'
 
     ezpublish.search.solr.document_mapper:
         alias: ezpublish.search.solr.document_mapper.native
