+++ lib/Resources/config/container/solr/document_field_mappers.yml
@@ -1,0 +1,67 @@
+parameters:
+    ezpublish.search.solr.document_field_mapper.block_translation.block_documents_content_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\ContentTranslation\BlockDocumentsContentFields
+    ezpublish.search.solr.document_field_mapper.block_translation.block_documents_meta_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\ContentTranslation\BlockDocumentsMetaFields
+    ezpublish.search.solr.document_field_mapper.content.content_document_base_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Content\ContentDocumentBaseFields
+    ezpublish.search.solr.document_field_mapper.content.content_document_location_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Content\ContentDocumentLocationFields
+    ezpublish.search.solr.document_field_mapper.content_translation.content_document_fulltext_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\ContentTranslation\ContentDocumentFulltextFields
+    ezpublish.search.solr.document_field_mapper.location.location_document_base_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location\LocationDocumentBaseFields
+    ezpublish.search.solr.document_field_mapper.location.location_document_content_fields.class: EzSystems\EzPlatformSolrSearchEngine\DocumentMapper\FieldMapper\Location\LocationDocumentContentFields
+
+services:
+    ezpublish.search.solr.document_field_mapper.block_translation.block_documents_content_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.block_translation.block_documents_content_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.content_type_handler'
+            - '@ezpublish.search.common.field_registry'
+            - '@ezpublish.search.common.field_name_generator'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.block_translation}
+
+    ezpublish.search.solr.document_field_mapper.block_translation.block_documents_meta_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.block_translation.block_documents_meta_fields.class%'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.block_translation}
+
+    ezpublish.search.solr.document_field_mapper.content.content_document_base_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.content.content_document_base_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.location_handler'
+            - '@ezpublish.spi.persistence.content_type_handler'
+            - '@ezpublish.spi.persistence.object_state_handler'
+            - '@ezpublish.spi.persistence.section_handler'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.content}
+
+    ezpublish.search.solr.document_field_mapper.content.content_document_location_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.content.content_document_location_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.location_handler'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.content}
+
+    ezpublish.search.solr.document_field_mapper.content_translation.content_document_fulltext_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.content_translation.content_document_fulltext_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.content_type_handler'
+            - '@ezpublish.search.common.field_registry'
+            - '@ezpublish.search.common.field_name_generator'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.content_translation}
+
+    ezpublish.search.solr.document_field_mapper.location.location_document_base_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.location.location_document_base_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.content_handler'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.location}
+
+    ezpublish.search.solr.document_field_mapper.location.location_document_content_fields:
+        class: '%ezpublish.search.solr.document_field_mapper.location.location_document_content_fields.class%'
+        arguments:
+            - '@ezpublish.spi.persistence.content_handler'
+            - '@ezpublish.spi.persistence.location_handler'
+            - '@ezpublish.spi.persistence.content_type_handler'
+            - '@ezpublish.spi.persistence.object_state_handler'
+            - '@ezpublish.spi.persistence.section_handler'
+        tags:
+            - {name: ezpublish.search.solr.document_field_mapper.location}
